Modified src/

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
-    public double calculate(double weight, double height){
+    public double calculate(double weight, double height) {
 
-        double bmi = (int)(weight/ (height * height));
+        int bmi = (int) (weight / (height * height));
         return bmi;
 
     }
