+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double weight = 52; //вес в килограммах
         double height = 1.60; // рост в метрах
-        double bmi = service.calculate(52, 1.60);
+        int bmi = (int) service.calculate(52, 1.60);
         System.out.println(bmi);
     }
 }
